// frontend/src/types/generated/DataSourceMetadata.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DataSourceTableInfo } from "./DataSourceTableInfo";

/**
 * Response containing all metadata about a connected data source
 */
export type DataSourceMetadata = { 
/**
 * The name of the data source
 */
name: string, 
/**
 * Schema information for all tables
 */
tables: { [key in string]: DataSourceTableInfo }, 
/**
 * First few rows of each table (table name -> rows)
 */
previewData: { [key in string]: Array<{ [key in string]: string }> }, };